requirements.txt
## Actual dependencies of the project
pure-python-adb==0.3.0.dev0
aiofiles==23.1.0
PyYAML==6.0.3
numpy~=2.4.3    # admits 2.4.x
Pillow~=12.1.1  # admits 12.1.x
# The dependencies will be upper bounded when it can be proven
# numpy>=2.4.3,<3.0 
# Pillow>=12.1.1,<13.0
